requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Vector database and embeddings
chromadb==0.4.18
sentence-transformers==2.2.2

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
pandas==2.1.3

# LLM integration
openai==1.3.7
anthropic==0.7.7

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
tenacity==8.2.3

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2